j2cl port

+++ packages/serverless-workflow-diagram-editor/kie-wb-common-stunner/kie-wb-common-stunner-client/kie-wb-common-stunner-widgets/src/main/java/org/kie/workbench/common/stunner/client/widgets/palette/categories/DefinitionPaletteCategoryWidget.java
@@ -24,8 +24,8 @@ import javax.annotation.PreDestroy;
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
 
-import org.jboss.errai.common.client.dom.HTMLElement;
-import org.jboss.errai.ioc.client.api.ManagedInstance;
+import elemental2.dom.HTMLElement;
+import io.crysknife.client.ManagedInstance;
 import org.kie.workbench.common.stunner.client.widgets.palette.categories.group.DefinitionPaletteGroupWidget;
 import org.kie.workbench.common.stunner.client.widgets.palette.categories.items.DefinitionPaletteItemWidget;
 import org.kie.workbench.common.stunner.core.client.components.palette.DefaultPaletteCategory;
@@ -75,10 +75,10 @@ public class DefinitionPaletteCategoryWidget implements DefinitionPaletteCategor
         this.itemMouseDownCallback = itemMouseDownCallback;
         final Glyph categoryGlyph = category.getGlyph();
         view.render(categoryGlyph,
-                    category.getIconSize(),
-                    category.getIconSize());
+                category.getIconSize(),
+                category.getIconSize());
         renderItems(category.getItems(),
-                    shapeFactory);
+                shapeFactory);
     }
 
     public void setOnOpenCallback(Consumer<DefaultPaletteCategory> onOpenCallback) {
@@ -108,13 +108,13 @@ public class DefinitionPaletteCategoryWidget implements DefinitionPaletteCategor
                 if (item instanceof PaletteGroup) {
 
                     renderGroup((DefaultPaletteGroup) item,
-                                shapeFactory);
+                            shapeFactory);
                 } else {
                     DefinitionPaletteItemWidget categoryItemWidget = definitionPaletteItemWidgetInstance.get();
 
                     categoryItemWidget.initialize(item,
-                                                  shapeFactory,
-                                                  itemMouseDownCallback);
+                            shapeFactory,
+                            itemMouseDownCallback);
 
                     view.addItem(categoryItemWidget);
                 }
@@ -127,8 +127,8 @@ public class DefinitionPaletteCategoryWidget implements DefinitionPaletteCategor
         DefinitionPaletteGroupWidget groupWidget = definitionPaletteGroupWidgetInstance.get();
 
         groupWidget.initialize(group,
-                               shapeFactory,
-                               itemMouseDownCallback);
+                shapeFactory,
+                itemMouseDownCallback);
 
         view.addGroup(groupWidget);
     }
@@ -143,16 +143,16 @@ public class DefinitionPaletteCategoryWidget implements DefinitionPaletteCategor
     }
 
     @Override
-    public void onMouseDown(int clientX,
-                            int clientY,
-                            int x,
-                            int y) {
+    public void onMouseDown(double clientX,
+                            double clientY,
+                            double x,
+                            double y) {
         if (itemMouseDownCallback != null) {
             itemMouseDownCallback.accept(new PaletteItemMouseEvent(category.getId(),
-                                                                   clientX,
-                                                                   clientY,
-                                                                   x,
-                                                                   y));
+                    clientX,
+                    clientY,
+                    x,
+                    y));
         }
     }
