+++ packages/serverless-workflow-diagram-editor/kie-wb-common-stunner/kie-wb-common-stunner-client/kie-wb-common-stunner-widgets/src/main/java/org/kie/workbench/common/stunner/client/widgets/palette/categories/DefinitionPaletteCategoryWidgetView.java
@@ -43,10 +43,10 @@ public interface DefinitionPaletteCategoryWidgetView extends UberElement<Definit
 
         DefaultPaletteCategory getCategory();
 
-        void onMouseDown(int clientX,
-                         int clientY,
-                         int x,
-                         int y);
+        void onMouseDown(double clientX,
+                         double clientY,
+                         double x,
+                         double y);
 
         void onClose();
