+++ packages/serverless-workflow-diagram-editor/kie-wb-common-stunner/kie-wb-common-stunner-client/kie-wb-common-stunner-widgets/src/main/java/org/kie/workbench/common/stunner/client/widgets/palette/categories/items/DefinitionPaletteItemWidget.java
@@ -23,7 +23,7 @@ import javax.annotation.PreDestroy;
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
 
-import org.jboss.errai.common.client.dom.HTMLElement;
+import elemental2.dom.HTMLElement;
 import org.kie.workbench.common.stunner.core.client.components.palette.AbstractPalette;
 import org.kie.workbench.common.stunner.core.client.components.palette.DefaultPaletteItem;
 import org.kie.workbench.common.stunner.core.client.components.palette.PaletteItemMouseEvent;
@@ -60,11 +60,11 @@ public class DefinitionPaletteItemWidget implements DefinitionPaletteItemWidgetV
                            Consumer<PaletteItemMouseEvent> itemMouseDownCallback) {
         this.item = item;
         final Glyph glyph = shapeFactory.getGlyph(item.getDefinitionId(),
-                                                  AbstractPalette.PaletteGlyphConsumer.class);
+                AbstractPalette.PaletteGlyphConsumer.class);
         this.itemMouseDownCallback = itemMouseDownCallback;
         view.render(glyph,
-                    item.getIconSize(),
-                    item.getIconSize());
+                item.getIconSize(),
+                item.getIconSize());
     }
 
     @Override
@@ -73,16 +73,17 @@ public class DefinitionPaletteItemWidget implements DefinitionPaletteItemWidgetV
     }
 
     @Override
-    public void onMouseDown(int clientX,
-                            int clientY,
-                            int x,
-                            int y) {
+    public void onMouseDown(double clientX,
+                            double clientY,
+                            double x,
+                            double y) {
         if (itemMouseDownCallback != null) {
+
             itemMouseDownCallback.accept(new PaletteItemMouseEvent(item.getId(),
-                                                                   clientX,
-                                                                   clientY,
-                                                                   x,
-                                                                   y));
+                    clientX,
+                    clientY,
+                    x,
+                    y));
         }
     }
