+++ packages/serverless-workflow-diagram-editor/kie-wb-common-stunner/kie-wb-common-stunner-client/kie-wb-common-stunner-widgets/src/main/java/org/kie/workbench/common/stunner/client/widgets/palette/collapsed/CollapsedDefinitionPaletteItemWidget.java
@@ -23,7 +23,7 @@ import javax.annotation.PreDestroy;
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
 
-import org.jboss.errai.common.client.dom.HTMLElement;
+import elemental2.dom.HTMLElement;
 import org.kie.workbench.common.stunner.core.client.components.palette.AbstractPalette;
 import org.kie.workbench.common.stunner.core.client.components.palette.CollapsedDefaultPaletteItem;
 import org.kie.workbench.common.stunner.core.client.components.palette.PaletteItemMouseEvent;
@@ -60,11 +60,11 @@ public class CollapsedDefinitionPaletteItemWidget implements CollapsedDefinition
                            final Consumer<PaletteItemMouseEvent> itemMouseDownCallback) {
         this.item = item;
         final Glyph glyph = shapeFactory.getGlyph(item.getDefinitionId(),
-                                                  AbstractPalette.PaletteGlyphConsumer.class);
+                AbstractPalette.PaletteGlyphConsumer.class);
         this.itemMouseDownCallback = itemMouseDownCallback;
         view.render(glyph,
-                    item.getIconSize(),
-                    item.getIconSize());
+                item.getIconSize(),
+                item.getIconSize());
     }
 
     @Override
@@ -73,16 +73,16 @@ public class CollapsedDefinitionPaletteItemWidget implements CollapsedDefinition
     }
 
     @Override
-    public void onMouseDown(final int clientX,
-                            final int clientY,
-                            final int x,
-                            final int y) {
+    public void onMouseDown(final double clientX,
+                            final double clientY,
+                            final double x,
+                            final double y) {
         if (itemMouseDownCallback != null) {
             itemMouseDownCallback.accept(new PaletteItemMouseEvent(item.getId(),
-                                                                   clientX,
-                                                                   clientY,
-                                                                   x,
-                                                                   y));
+                    clientX,
+                    clientY,
+                    x,
+                    y));
         }
     }
