+++ packages/serverless-workflow-diagram-editor/kie-wb-common-stunner/kie-wb-common-stunner-client/kie-wb-common-stunner-widgets/src/main/java/org/kie/workbench/common/stunner/client/widgets/palette/collapsed/CollapsedDefinitionPaletteItemWidgetView.java
@@ -31,9 +31,9 @@ public interface CollapsedDefinitionPaletteItemWidgetView extends UberElement<Co
 
         CollapsedDefaultPaletteItem getItem();
 
-        void onMouseDown(final int clientX,
-                         final int clientY,
-                         final int x,
-                         final int y);
+        void onMouseDown(final double clientX,
+                         final double clientY,
+                         final double x,
+                         final double y);
     }
 }
